requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.20
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
